+++ test/ex_litedb_test.exs
@@ -1,0 +1,4 @@
+defmodule ExLitedbTest do
+  use ExUnit.Case
+  doctest ExLitedb
+end