+++ test/ex_sqlean_test.exs
@@ -1,4 +1,0 @@
-defmodule ExSqleanTest do
-  use ExUnit.Case
-  doctest ExSqlean
-end