+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.9.0
+
+- Rename package to `ex_litedb`
+-
+
 ### 0.8.8
 
 - Update to sqlean 0.8.8-3360 (https://github.com/mindreframer/sqlean/releases/tag/0.8.8-3360)
@@ -7,7 +12,7 @@
 
 - Update to sqlean 0.8.7-3360 (https://github.com/mindreframer/sqlean/releases/tag/0.8.7-3360)
 - https://github.com/mindreframer/sqlean/blob/forked/CHANGELOG.md#087-3360
-- make extensions list in `ExSqlean.extensions()` dynamic
+- make extensions list in `ExLitedb.extensions()` dynamic
 
 ### 0.8.6